problem with var
since times>3 is true, greeter is redefined to "say hello instead".

While this is not a problem if you knowingly want greeter to be redefined, it becomes a problem when you do not realize that a variable greeter has already been defined before.

If you have used greeter in other parts of your code, you might be surprised at the output you might get. This will likely cause a lot of bugs in your code.

 This is why let and const are necessary.